getElementsByClassName is now composed of two recursive functions:
nodeCheckerRecursiveRouter:  takes a node and checks if it matches the class.  If it does it adds it to desiredElements.  Then it determines if the node has children.  If it does it passes the node to childNodeTraverser.  If it does not it passes the first node from remainingNodes to childNodeTraverser.

childNodeTraverser:  finds a node's children, adding the extra nodes to remainingNodes.  Then it passes the first child to nodeCheckerRecursiveRouter.

childNodeTraverser doesn't need to exist, but I like it for readability.

Current bug:  we keep throwing an error when we hit line 52.  matchesClass is getting undefined passed to it somehow.  the functions are all operative, but there's an issue with how data is being stored/accessed.